requirements.txt

### REQUIREMENTS

Need have 2 ambients:
    1) Prod
        -> Need put one superuser
    2) stg
        -> I can put others user, for use app

In the sideBar:
    Dashboard:
        CRUD my revenues per monthy
        CRUD my costs per monthy

    Input data:

### FOR CREATE A AUTH SYSTEM

For create a authentication system with django

in the HTML I need have:

<form action="/login/" method="post">
    {% csrf_token %}
    <div id="login">
        <label for="username">Username</label>
        <input type="text" name="username" id="input_username">
    </div>
    <div id="password">
        <label for="password">Password</label>
        <input type="password" name="password" id="input_password">
    </div>
    <div id="enter">
        <button type="submit" id="button_enter">Enter</button>
    </div>
</form>

in views.py on the my app:

1) Import libs important for auth:

from django.http import HttpResponse, HttpResponseRedirect
from django.shortcuts import render
from django.contrib.auth import authenticate, login as django_login, logout as django_logout
from django.contrib.auth.decorators import login_required
from django.shortcuts import redirect, get_object_or_404

2) Create the method for validation users:

def indexAuth(request):
    if request.method == 'GET':
        return render(request, 'index.html')
    
    #Se não o metodo é POST
    
    username = request.POST.get('username')
    password = request.POST.get('password')

    user = authenticate(username=username, password=password)
    
    if user:
        django_login(request, user)
        return HttpResponseRedirect('/home/') 
    
    return render(request, 'index.html')

3) Remember: 

You need create the route for make the redirect when user to be valitation.



